task7/requirements.txt
streamlit
yfinance
forex-python
pandas
feedparser
plotly
python-dotenv